+++ jero-core/src/main/java/com/lxt/core/enums/DeleteEnum.java
@@ -2,18 +2,18 @@ package com.lxt.core.enums;
 
 public enum DeleteEnum {
 
-    NORMAL(1, "正常"),
-    DELETE(0, "已删除");
+    NORMAL((short)1, "正常"),
+    DELETE((short)0, "已删除");
 
-    private int value;
+    private short value;
     private String label;
 
-    DeleteEnum(int value, String label){
+    DeleteEnum(short value, String label){
         this.value = value;
         this.label = label;
     }
 
-    public int getValue(){return value;}
+    public short getValue(){return value;}
     public String getＬabel(){return label;}
 
 }
